requirements.txt
Django~=5.1.0
django-stubs-ext~=5.1.0
django-filter~=24.3
djangorestframework~=3.15.0
drf-spectacular~=0.27.2
python-dotenv~=1.0.1
django-extensions~=3.2.3
pygraphviz~=1.14